+++ index.js
@@ -2,8 +2,8 @@ require('dotenv').config();
 
 const grpc = require('grpc');
 const protoLoader = require('@grpc/proto-loader');
-const { resolvedUrls } = require('./src/resolvers/resolvedUrls');
-const { getBrowserStats } = require('./src/resolvers/browserStats');
+const { resolveUrls } = require('./src/resolvers/resolveUrls');
+const { getBrowserStats } = require('./src/resolvers/browser');
 
 const PROTO_PATHS = {
   urlResolver: __dirname + '/src/typeDefs/url_resolver.proto',
@@ -35,7 +35,7 @@ const PORT = process.env.PORT || 4000;
 
 const server = new grpc.Server();
 server.addService(urlResolverProto.UrlResolver.service, {
-  resolveUrl: resolvedUrls,
+  ResolveUrl: resolveUrls,
 });
 server.addService(browserProto.BrowserStats.service, {
   GetStats: getBrowserStats,
