+++ src/resolvers/browser.js
